requirements.txt
pytest==6.2.4
requests==2.26.0
pandas==1.3.0
PyYAML==5.4.1
